requirements.txt
amqp==5.0.6
billiard==3.6.4.0
celery==5.2.2
kombu==5.2.2
msgpack==1.0.3
Pillow==9.3.0
gcloud==0.18.3
Django==4.0.9
django-storages[google]==1.12.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
django-filter==21.1
django-enumfields==2.1.1
django-healthz==0.0.5
django-rehive-extras==1.1.2
drf-rehive-extras==1.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
requests==2.26.0
psycopg2-binary==2.9.3
rehive==1.2.19
sentry-sdk==1.5.6
vine==5.0.0
